pyproject.toml
[project]
name = "mini-autentificacao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "uvicorn>=0.41.0",
]
